requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.6.0
gunicorn>=20.1.0
joblib>=1.2.0
requests>=2.28.0
python-dotenv
groq>=0.4.0
pymongo[srv]>=4.0.0
dnspython>=2.2.0
bcrypt>=4.0.0
reportlab>=4.0.0
certifi>=2024.0.0
